refactor(browser): split actScript and ai_policy for the file-size gate
check:all's file-size gate caught both files over 300 lines after the
WI-NB1/NB3 additions. Split along the natural seams: the injected ES5
library moves to agentLib.ts (actScript keeps the builders and re-exports
AGENT_LIB for its sibling importers), and the pure address math -
blocked ranges, transition-prefix IPv4 extraction, legacy IPv4
spellings - moves to ai_policy_addr.rs (ai_policy keeps URL/hostname
policy and blocked_ip). No behaviour change; agentLib.test.ts pins the
library assembly so a dropped concat member cannot ship as a page-side
ReferenceError.
