+++ src-tauri/src/browser/mod.rs
@@ -41,6 +41,7 @@
 pub mod ai_commands;
 pub mod ai_guards;
 pub mod ai_policy;
+pub mod ai_policy_addr;
 pub mod authorize;
 pub mod commands;
 pub mod commands_auth;
